refactor(api): festival-info → src/api modules (Effort 1)
Move the festival-info data-access slice from src/hooks/queries/festival-info
to src/api/festival-info per ADR 0001. Adds a shared types.ts (FestivalInfo Row
type + query-key factory) and a per-file festivalInfoQuery queryOptions factory.
No behavior change; consumers only change import paths.
