+++ src/api/festival-info/useFestivalInfoMutation.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { supabase } from "@/integrations/supabase/client";
 import { Tables } from "@/integrations/supabase/types";
-import { festivalInfoKeys } from "./useFestivalInfo";
+import { festivalInfoKeys } from "./types";
 
 export function useFestivalInfoMutation(festivalId: string) {
   const queryClient = useQueryClient();
