+++ src/pages/EditionView/tabs/InfoTab.tsx
@@ -1,5 +1,5 @@
 import { useFestivalEdition } from "@/contexts/FestivalEditionContext";
-import { useFestivalInfoQuery } from "@/hooks/queries/festival-info/useFestivalInfo";
+import { useFestivalInfoQuery } from "@/api/festival-info/useFestivalInfo";
 import { PageTitle } from "@/components/PageTitle/PageTitle";
 import { EditionTitle } from "./InfoTab/EditionTitle";
 import { InfoText } from "./InfoTab/InfoText";
